ms-audit/requirements.txt
# gRPC
grpcio==1.60.0
grpcio-tools==1.60.0
protobuf==4.25.1

# Database
psycopg2-binary==2.9.9

# Redis
redis==5.0.1

# Metrics
prometheus-client==0.19.0

# Logging
python-json-logger==2.0.7

# Utils
python-dotenv==1.0.0

